pyproject.toml
[project]
name = "scalable-search-agent"
version = "0.1.0"
description = "LLM-powered Internet-Search Agent for financial research"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-ollama>=0.2.0",
    "tavily-python>=0.5.0",
    "pydantic-settings>=2.7.0",
    "sse-starlette>=2.2.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
    "python-semantic-release>=9.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.semantic_release]
version_variable = "pyproject.toml:version"
branch = "main"
commit_message = "chore(release): {version}"
build_command = ""

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
